Move leadingZeroes check to Tx/util.ts
