+++ packages/tx/src/eip1559Transaction.ts
@@ -3,10 +3,8 @@ import {
   bnToHex,
   bnToUnpaddedBuffer,
   ecrecover,
-  hasLeadingZeroes,
   keccak256,
   rlp,
-  RlpValues,
   toBuffer,
 } from 'ethereumjs-util'
 import Common from '@ethereumjs/common'
@@ -19,8 +17,9 @@ import {
   JsonTx,
   N_DIV_2,
   TxOptions,
+  RlpValues,
 } from './types'
-import { AccessLists } from './util'
+import { AccessLists, hasLeadingZeroes } from './util'
 
 const TRANSACTION_TYPE = 2
 const TRANSACTION_TYPE_BUFFER = Buffer.from(TRANSACTION_TYPE.toString(16).padStart(2, '0'), 'hex')
