+++ packages/util/src/bytes.ts
@@ -1,6 +1,6 @@
 import { BN } from './externals'
 import { stripHexPrefix, padToEven, isHexString, isHexPrefixed } from './internal'
-import { PrefixedHexString, RlpValues, TransformableToArray, TransformableToBuffer } from './types'
+import { PrefixedHexString, TransformableToArray, TransformableToBuffer } from './types'
 import { assertIsBuffer, assertIsArray, assertIsHexString } from './helpers'
 
 /**
@@ -275,16 +275,4 @@ export const baToJSON = function (ba: any): any {
     }
     return array
   }
-}
-
-/**
- *
- * @param values An object containing a set keys and RLP encoded values
- * @returns a string with the value that was encoded with leading zeroes or undefined if no values have leading zeroes
- */
-export const hasLeadingZeroes = function (values: RlpValues): string | undefined {
-  const res = Object.entries(values).find(
-    (entry) => entry[1] && entry[1]?.length > 0 && entry[1][0] === 0x00
-  )
-  return res ? res[0] : undefined
 }
